// src/constants/enums.ts
export enum WfeButtonType {
  ChangePassword = 'ChangePassword',
  ChangeUsername = 'ChangeUsername',
  EnableMFA = 'EnableMFA',
  DisableMFA = 'DisableMFA'
}

export enum StepChoice {
  submit = 'submit',
  resend = 'resend',
  yes = 'Yes',
  no = 'No',
  ok = 'ok',
  retry = 'retry',
  confirm = 'Confirm',
  keep = 'Keep'
}

export enum AuthType {
  authenticated = 'authenticated',
  unauthenticated = 'unauthenticated'
}